Build: Add Linting and prefetch Astro integration
* Build: Add Linting and prefetch Astro integration

TODO: tsx files currently have the problem that ESLint reports two errors that need to be fixed in the future:
* `Parsing error: "parserOptions.project" has been set for @typescript-eslint/parser.
The file does not match your project config: src/**/abc.tsx.
The file must be included in at least one of the projects provided`
* `'React' refers to a UMD global, but the current file is a module. Consider adding an import instead.`
Reason for the second error: It is not recognized that JSX Transform (https://reactjs.org/blog/2020/09/22/introducing-the-new-jsx-transform.html) is used.